refactor: move filter functions to separate file
